'subtotal with discounted apples in basket contains 3 apples & an orange' implementation

+++ src/main/scala/Discount.scala
@@ -1,14 +1,14 @@
 sealed abstract class Discount(itemName: String) {
-  def calculate(basket: Basket, basePrices: Map[String, Double] = Map()): Double
+  def calculate(basket: Basket, basePrices: Map[String, BigDecimal] = Map()): BigDecimal
 
   protected def calculateByOccurance(
     basket: Basket,
-    basePrices: Map[String, Double],
+    basePrices: Map[String, BigDecimal],
     occurance: Int
-  ): Double = {
+  ): BigDecimal = {
     val filteredItems = basket.items.filter(_ == itemName)
     if (filteredItems.isEmpty) {
-      0d
+      BigDecimal(0)
     } else {
       val numberOfElements  = filteredItems.size
       val discountOccurance = numberOfElements / occurance
@@ -17,12 +17,12 @@ sealed abstract class Discount(itemName: String) {
   }
 }
 case class ThreeForTwoDiscount(itemName: String) extends Discount(itemName) {
-  override def calculate(basket: Basket, basePrices: Map[String, Double]): Double = {
+  override def calculate(basket: Basket, basePrices: Map[String, BigDecimal]): BigDecimal = {
     calculateByOccurance(basket, basePrices, 3)
   }
 }
 case class BuyOneGetOneDiscount(itemName: String) extends Discount(itemName) {
-  override def calculate(basket: Basket, basePrices: Map[String, Double] = Map()): Double = {
+  override def calculate(basket: Basket, basePrices: Map[String, BigDecimal] = Map()): BigDecimal = {
     calculateByOccurance(basket, basePrices, 2)
   }
 }
